4-geminiai-integration/requirements.txt
mcp[cli]==1.10.1
google-generativeai
python-dotenv
ipykernel
httpx
